DONE!
xxhash32/64 should be done. The full test suite & benchmarking utility
is added. Adding all the tests actually flushed out a bug within XXHash64
I am extremely happy to see that get caught.

To future implementors: It is easy to skip small inputs which do buffer
stuff, and the finalize math is kind of _weird_ so be sure to test small
un-even inputs.